*updated notebooks, readme *moved class desc to base.py *Added Runtime Enum
