+++ QEfficient/src/base.py
@@ -1,5 +1,52 @@
+# -----------------------------------------------------------------------------
+#
+# Copyright (c) 2024 Qualcomm Innovation Center, Inc. All rights reserved.
+# SPDX-License-Identifier: BSD-3-Clause
+#
+# ----------------------------------------------------------------------------
+
+"""
+** This file for holds the classes that handle main functions
+1.load i.e. from_pretrained
+2.execute
+3.transform
+4.export
+5.compile
+For different varities of Transformer Models
+
+Representation of class inheritence followed keeping in line with transformers/diffusers repos ->
+
+                                                                                            QEFFBaseModel
+                                                 ________________________________________________|________________________________________________________________
+                                                |                                                                                                                 |  
+                                            QEFFTransformersBase                                                                                           QEFFDiffusersBase
+                                                |                                                                                                                 |
+                                    ____________|________________________________________________________ ________________                       _________________|______________
+                   _____           |                              |                                      |                |                     |                                |         
+                  |          QEFFAutoModel             QEFFAutoModelForCausalLM              QEFFAWQModelForCausalLM     ...                   ...                              ...
+QEFFCommonLoader -|       [Provides way to          [Provides way to do 1-5 on                 [Supports 1-5 for 
+[Provides         |        do steps 1-5 on           transformers.AutoModelForCausalLM]         AWQ Models]
+interface to      |_____   transformers.AutoModel]
+Load any of 
+These models       
+by automatically
+detecting the type
+of the model]
+
+** QEFFBASEModel is abstract base class that defines the basic structure of these classes.
+** QEFFPipeline classes will stay at the same level as QEFFAutoModel in this hierarchy in future.
+"""
+
 from abc import ABC, abstractmethod
+from enum import Enum
 from typing import Any
+
+
+#Defining placeholder ENUM for execute function
+class Runtime(Enum):
+    CPU_ORT = "CPU ONNX Runtime"
+    CPU_PT = "CPU PyTorch Runtime"
+    AI_100 = "AI_100"
 
 
 class QEFFBaseModel(ABC):
@@ -13,6 +60,11 @@ class QEFFBaseModel(ABC):
         super().__init__()
         # Users can call generate or execute
         self.generate = self.execute
+        self._runtime = Runtime.CPU_PT
+
+    @property
+    def runtime(self) -> Runtime:
+        return self._runtime
 
     @classmethod
     def from_pretrained(cls, pretrained_model_name_or_path: str, *args, **kwargs):
